feat: add PendingRequests component and integrate into CodeConnectDetails
