+++ frontend/src/types/codeConnectTypes.ts
@@ -31,7 +31,6 @@ export interface ApiProjectContributor {
 
 export interface ApiProjectData {
   id: number;
-  user_id: number;
   contributors: ApiProjectContributor[];
   description?: string;
   language_backend: string;
@@ -39,18 +38,6 @@ export interface ApiProjectData {
   roadmap?: { task: string; done: boolean }[];
   time_duration: string;
   title: string;
-}
-
-export interface ApiContributor {
-  id: number;
-  user_id: number;
-  programming_role: string;
-  status: "pending" | "accepted" | "rejected";
-  user: {
-    id: number;
-    name: string;
-    email: string;
-  };
 }
 
 export interface ApiProjectsResponse {
